feat: Provider周りをリファクタ
